requirements.txt
streamlit
streamlit-aggrid
streamlit-extras
pandas
supabase
python-dotenv
